Move GFX950 MoE GEMM tests to AMD package
